Removed the use of float in favor of double to prevent loss of precision.

+++ src/ElasticSearch.Client/Domain/Facets/DateRangeFacet.cs
@@ -21,19 +21,19 @@ namespace ElasticSearch.Client
             public DateTime? From { get; internal set; }
 
             [JsonProperty(PropertyName = "min")]
-            public float? Min { get; internal set; }
+            public double? Min { get; internal set; }
 
             [JsonProperty(PropertyName = "max")]
-            public float? Max { get; internal set; }
+            public double? Max { get; internal set; }
 
             [JsonProperty(PropertyName = "total_count")]
             public int TotalCount { get; internal set; }
 
             [JsonProperty(PropertyName = "total")]
-            public float Total { get; internal set; }
+            public double Total { get; internal set; }
 
             [JsonProperty(PropertyName = "mean")]
-            public float? Mean { get; internal set; }
+            public double? Mean { get; internal set; }
         }
 
         #endregion
